Move theme to utils folder, components, change bg to black
